backend/requirements.txt
fastapi
uvicorn
qdrant-client
python-dotenv
pydantic
python-multipart
requests
PyPDF2
python-docx
